change class JavaIOPostRepositoryImpl and JavaIOUserRepositoryImpl

+++ src/main/java/com/ovani4/crudprogram/repository/PostRepository.java
@@ -2,6 +2,6 @@ package main.java.com.ovani4.crudprogram.repository;
 
 import main.java.com.ovani4.crudprogram.model.Post;
 
-public interface PostRepository extends GenericRepository <Post, Long>{
-    //методы для создания и редактирование post.txt
+public interface PostRepository extends GenericRepository <Post, Integer>{
+
 }
