+++ src/main/java/com/ovani4/crudprogram/repository/UserRepository.java
@@ -2,6 +2,6 @@ package main.java.com.ovani4.crudprogram.repository;
 
 import main.java.com.ovani4.crudprogram.model.User;
 
-public interface UserRepository extends GenericRepository <User, Long>{
-    //методы для создания и редактирование user.txt
+public interface UserRepository extends GenericRepository <User, Integer>{
+
 }
